Split out arrow-integration-test crate
