+++ arrow-integration-test/Cargo.toml
@@ -1,0 +1,48 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+[package]
+name = "arrow-integration-test"
+version = "24.0.0"
+description = "Support for the Apache Arrow JSON test data format"
+homepage = "https://github.com/apache/arrow-rs"
+repository = "https://github.com/apache/arrow-rs"
+authors = ["Apache Arrow <dev@arrow.apache.org>"]
+license = "Apache-2.0"
+keywords = ["arrow"]
+include = [
+    "benches/*.rs",
+    "src/**/*.rs",
+    "Cargo.toml",
+]
+edition = "2021"
+rust-version = "1.62"
+
+[lib]
+name = "arrow_integration_test"
+path = "src/lib.rs"
+bench = false
+
+[dependencies]
+arrow = { version = "24.0.0", path = "../arrow", default-features = false }
+arrow-buffer = { version = "24.0.0", path = "../arrow-buffer" }
+hex = { version = "0.4", default-features = false, features = ["std"] }
+serde = { version = "1.0", default-features = false, features = ["rc", "derive"] }
+serde_json = { version = "1.0", default-features = false, features = ["std"] }
+num = { version = "0.4", default-features = false, features = ["std"] }
+
+[build-dependencies]
