+++ arrow-integration-testing/src/bin/arrow-file-to-stream.rs
