+++ arrow-integration-testing/src/bin/arrow-json-integration-test.rs
@@ -20,7 +20,8 @@ use arrow::datatypes::{DataType, Field};
 use arrow::error::{ArrowError, Result};
 use arrow::ipc::reader::FileReader;
 use arrow::ipc::writer::FileWriter;
-use arrow_integration_testing::{read_json_file, util::*};
+use arrow_integration_test::*;
+use arrow_integration_testing::read_json_file;
 use clap::Parser;
 use std::fs::File;
