+++ arrow-integration-testing/src/bin/arrow-stream-to-file.rs
