+++ arrow-integration-testing/src/flight_client_scenarios/auth_basic_proto.rs
