+++ arrow-integration-testing/tests/ipc_reader.rs
